1.商品待申请批量上传图片 逻辑修改 2.商品进销存信息,交易倍数修改

+++ src/main/java/com/aiqin/bms/scmp/api/product/domain/pojo/ProductSkuDisInfoDraft.java
@@ -50,5 +50,5 @@ public class ProductSkuDisInfoDraft extends CommonBean {
     private String productSkuName;
 
     @ApiModelProperty("拆零系数")
-    private BigDecimal zeroRemovalCoefficient;
+    private Long zeroRemovalCoefficient;
 }
