+++ src/main/java/com/aiqin/bms/scmp/api/product/domain/pojo/ProductSkuSalesInfoDraft.java
@@ -66,5 +66,5 @@ public class ProductSkuSalesInfoDraft extends CommonBean {
     private String productSkuName;
 
     @ApiModelProperty("拆零系数")
-    private BigDecimal zeroRemovalCoefficient;
+    private Long zeroRemovalCoefficient;
 }
